Migrate to an src/ content layout
The amazing thing about this change is that it greatly clarifies the
separation between site source files (everything in src/) and
configuration/support files (everything not in src/). This helps avoid
accidentally deploying files that we don't want to deploy. It also
clarifies how the "static" directory is not copied-as-is in the HTML,
and should not be linked to directly; instead static is an input that
sphinx builds into the _static output dir.