Add env-dependent account and log in docs
This is a reworking of the logging-in.rst documentation to implement RSP
environment-specific documentation builds.

Information about getting an account is now split out into account.rst.
This is where we list GitHub teams, Slack channels to request access,
link back to the main docs for internal environments, and discuss any
VPN installation.

This simplifies the logging-in docs so we can concentrate on literally
logging in and configuring the machine.

The templating here is done using sphinx-jinja and data in
rsp-environments.yaml.