Add rsplint.py linter
This linter ensures that data in rsp-environments.yaml matches what's in
https://github.com/lsst-sqre/phalanx. In particular, it checks that URLs
match what's expected from the fqdn, and that GitHub teams configured
for Gafaelfawr auth are consistent.

This also adds infrastructure to test the quality of the linter's code
(black, isort, flake8 and mypy). To install dependencies for this
"development" infrastructure, I've added the requirements/dev.in file.